+++ ui/src/components/footer.vue
@@ -1,8 +1,8 @@
 <template lang="pug">
 div
   .footer.w-100.flex-wrap.align-content-center(style="display:none; opacity: 0;")
-    .footer-inner.container.d-flex.flex-wrap.align-content-center
-      .col.p-0.m-0 © SeydX
+    .footer-inner.container.d-flex.flex-wrap.align-content-center.toggleArea
+      .col.p-0.m-0.toggleArea © SeydX
       a(href="https://www.github.com/SeydX/homebridge-camera-ui", target="_blank" rel="noopener noreferrer")
         b-icon.text-color-primary(icon="github", aria-hidden="true")
   .snapline